requirements.txt
streamlit
pytesseract
Pillow
beautifulsoup4
requests
